+++ nest-forms-backend/src/forms/guards/form-must-be-enabled.guard.ts
@@ -6,11 +6,13 @@ import {
   Injectable,
   NotFoundException,
 } from '@nestjs/common'
+import { Reflector } from '@nestjs/core'
 import { Request } from 'express'
 import { getFormDefinitionBySlug } from 'forms-shared/definitions/getFormDefinitionBySlug'
-import PrismaService from 'src/prisma/prisma.service'
 
 import { User } from '../../auth-v2/types/user'
+import { ALLOW_COMPLETED_DISABLED_FORMS_KEY } from '../../forms-v2/decorators/allow-completed-disabled-forms.decorator'
+import FormsService from '../forms.service'
 
 interface RequestWithUser extends Request {
   user?: User
@@ -18,7 +20,10 @@ interface RequestWithUser extends Request {
 
 @Injectable()
 export class FormMustBeEnabledGuard implements CanActivate {
-  constructor(private readonly prismaService: PrismaService) {}
+  constructor(
+    private readonly formsService: FormsService,
+    private readonly reflector: Reflector,
+  ) {}
 
   async canActivate(context: ExecutionContext): Promise<boolean> {
     const request = context.switchToHttp().getRequest<RequestWithUser>()
@@ -29,22 +34,25 @@ export class FormMustBeEnabledGuard implements CanActivate {
       throw new BadRequestException('formId path parameter is required')
     }
 
-    const form = await this.prismaService.forms.findUnique({
-      where: { id: formId },
-    })
-    if (!form) {
-      // TODO: Errors
-      throw new NotFoundException('form not found')
-    }
+    const form = await this.formsService.getForm(formId)
 
     const formDefinition = getFormDefinitionBySlug(form.formDefinitionSlug)
     if (!formDefinition) {
       // TODO: Errors
       throw new NotFoundException('form definition not found')
     }
+
     if (formDefinition.isDisabled) {
-      // TODO: Errors
-      throw new ForbiddenException('form definition is disabled')
+      const allowCompletedDisabledForms =
+        this.reflector.getAllAndOverride<boolean>(
+          ALLOW_COMPLETED_DISABLED_FORMS_KEY,
+          [context.getHandler(), context.getClass()],
+        )
+
+      if (!allowCompletedDisabledForms || this.formsService.isEditable(form)) {
+        // TODO: Errors
+        throw new ForbiddenException('form definition is disabled')
+      }
     }
 
     return true
